Day 15 part 1
I got really bogged down in this.

1. It's quite a long problem description.
2. I got lost.

I was stuck for a while trying to get A* to work in the implementation
of unit.nextStep.

I thought it would be better to start with the list of enemies, and use
the heuristic of the Manhattan distance. That gives the lowest
theoretical cost to get to a neighbouring square.

```python
closest_distance = MAX_INT
next_step = None
target = None

for enemy in ordered_enemies:

  if manhattan_distance(unit, enemy) > closest_distance:
    # this enemy and all others in the ordered list have a theoretical
    # best distance exceeding the current path to the enemy.
    # Stop evaluating any other options.
    return (next_step, target)
  distances, paths = astar_search(unit, enemy)
  switch distances[unit]:
    case < closest_distance:
      next_step = first_step(paths, paths[enemy], u)
      target = paths[enemy]
    case == closest_distance:
      # should already have lost since the enemies are in reading order?
      target = reading_order(target, paths[enemy])[0]

return (next_step, target)
```

I couldn't get that working. The reading order thing was tripping me up.

I ended up looking at a solution which gave me a different approach in
nextStep.

See https://bitbucket.org/thezeez/advent-of-code-2018