Day 19 part 2
This was a bit invovled. I got Day 19 part 1 reasonably quickly
mechanically. But it does run really slow (a minuite or so) with logging
output as per the example. So turn that off.

The second part needed analysis of the actual program being interpreted.

It's summing the factors of the contents of register 3 (zero-based
index). So wrote a shortcut which detects when that part of the program
has been entered (from instruction 1 onwards) and does it via idiomatic
Go, rather than the interpreter.

That required reading the assembly. See the onnotated version of the
input for more details.